refactor : dto 변경
